# fix hisat2 performance: remove unused SpliceSiteDB from AlnSink