@yttria-aniseia 2024-08-07
hm, no, this is wrong...  hisat2 does use AlnSink's SpliceSiteDB, it's output to the `novelSpliceSiteOutfile` when provided.
this is the `--no-temp-splicesite` again.  
but the structure itself is maintained even without `novelSpliceSiteOutfile` and `knownSpliceSiteInfile` aren't provided, because 
```
bool write = novelSpliceSiteOutfile != "" || useTempSpliceSite;
bool read = knownSpliceSiteInfile != "" || novelSpliceSiteInfile != "" || useTempSpliceSite;
```